マイグレーション修正

+++ db/migrate/20230622035924_drop_rooms.rb
@@ -1,5 +1,5 @@
 class DropRooms < ActiveRecord::Migration[6.0]
   def change
-    drop_table :rooms
+    drop_table :rooms if table_exists?(:rooms)
   end
 end
